+++ packages/core/src/agent/index.ts
@@ -15,10 +15,9 @@ import type {
   UserContent,
 } from 'ai';
 import type { JSONSchema7 } from 'json-schema';
-import type { ZodSchema } from 'zod';
-import { z } from 'zod';
+import type { ZodSchema, z } from 'zod';
 
-import type { MastraPrimitives } from '../action';
+import type { MastraPrimitives, MastraUnion } from '../action';
 import { MastraBase } from '../base';
 import type { Metric } from '../eval';
 import { AvailableHooks, executeHook } from '../hooks';
@@ -26,13 +25,15 @@ import type { GenerateReturn, StreamReturn } from '../llm';
 import type { MastraLLMBase } from '../llm/model';
 import { MastraLLM } from '../llm/model';
 import { RegisteredLogger } from '../logger';
+import type { Mastra } from '../mastra';
 import type { MastraMemory } from '../memory/memory';
 import type { MemoryConfig, StorageThreadType } from '../memory/types';
 import { InstrumentClass } from '../telemetry';
-import type { CoreTool, ToolAction } from '../tools/types';
+import type { CoreTool } from '../tools/types';
+import { makeCoreTool, createMastraProxy } from '../utils';
 import type { CompositeVoice } from '../voice';
 
-import type { AgentConfig, AgentGenerateOptions, AgentStreamOptions, ToolsetsInput } from './types';
+import type { AgentConfig, AgentGenerateOptions, AgentStreamOptions, ToolsetsInput, ToolsInput } from './types';
 
 export * from './types';
 
@@ -41,14 +42,14 @@ export * from './types';
   excludeMethods: ['__setTools', '__setLogger', '__setTelemetry', 'log'],
 })
 export class Agent<
-  TTools extends Record<string, ToolAction<any, any, any, any>> = Record<string, ToolAction<any, any, any, any>>,
+  TTools extends ToolsInput = ToolsInput,
   TMetrics extends Record<string, Metric> = Record<string, Metric>,
 > extends MastraBase {
   public name: string;
   readonly llm: MastraLLMBase;
   instructions: string;
   readonly model?: LanguageModelV1;
-  #mastra?: MastraPrimitives;
+  #mastra?: Mastra;
   #memory?: MastraMemory;
   tools: TTools;
   /** @deprecated This property is deprecated. Use evals instead. */
@@ -78,7 +79,10 @@ export class Agent<
     }
 
     if (config.mastra) {
-      this.#mastra = config.mastra;
+      this.__registerPrimitives({
+        telemetry: config.mastra.getTelemetry(),
+        logger: config.mastra.getLogger(),
+      });
     }
 
     if (config.metrics) {
@@ -123,9 +127,11 @@ export class Agent<
 
     this.llm.__registerPrimitives(p);
 
-    this.#mastra = p;
-
     this.logger.debug(`[Agents:${this.name}] initialized.`, { model: this.model, name: this.name });
+  }
+
+  __registerMastra(mastra: Mastra) {
+    this.#mastra = mastra;
   }
 
   /**
@@ -138,7 +144,8 @@ export class Agent<
   }
 
   async generateTitleFromUserMessage({ message }: { message: CoreUserMessage }) {
-    const { object } = await this.llm.__textObject<{ title: string }>({
+    // need to use text, not object output or it will error for models that don't support structured output (eg Deepseek R1)
+    const { text } = await this.llm.__text<{ title: string }>({
       messages: [
         {
           role: 'system',
@@ -146,19 +153,19 @@ export class Agent<
       - you will generate a short title based on the first message a user begins a conversation with
       - ensure it is not more than 80 characters long
       - the title should be a summary of the user's message
-      - do not use quotes or colons`,
+      - do not use quotes or colons
+      - the entire text you return will be used as the title`,
         },
         {
           role: 'user',
           content: JSON.stringify(message),
         },
       ],
-      structuredOutput: z.object({
-        title: z.string(),
-      }),
     });
 
-    return object.title;
+    // Strip out any r1 think tags if present
+    const cleanedText = text.replace(/<think>[\s\S]*?<\/think>/g, '').trim();
+    return cleanedText;
   }
 
   getMostRecentUserMessage(messages: Array<CoreMessage>) {
@@ -167,7 +174,7 @@ export class Agent<
   }
 
   async genTitle(userMessage: CoreUserMessage | undefined) {
-    let title = 'New Thread';
+    let title = `New Thread ${new Date().toISOString()}`;
     try {
       if (userMessage) {
         title = await this.generateTitleFromUserMessage({
@@ -198,18 +205,20 @@ export class Agent<
     const userMessage = this.getMostRecentUserMessage(userMessages);
     const memory = this.getMemory();
     if (memory) {
+      const config = memory.getMergedThreadConfig(memoryConfig);
       let thread: StorageThreadType | null;
+
       if (!threadId) {
         this.logger.debug(`No threadId, creating new thread for agent ${this.name}`, {
           runId: runId || this.name,
         });
-        const title = await this.genTitle(userMessage);
+        const title = config?.threads?.generateTitle ? await this.genTitle(userMessage) : undefined;
 
         thread = await memory.createThread({
           threadId,
           resourceId,
-          title,
           memoryConfig,
+          title,
         });
       } else {
         thread = await memory.getThreadById({ threadId });
@@ -217,7 +226,9 @@ export class Agent<
           this.logger.debug(`Thread with id ${threadId} not found, creating new thread for agent ${this.name}`, {
             runId: runId || this.name,
           });
-          const title = await this.genTitle(userMessage);
+
+          const title = config?.threads?.generateTitle ? await this.genTitle(userMessage) : undefined;
+
           thread = await memory.createThread({
             threadId,
             resourceId,
@@ -247,6 +258,7 @@ export class Agent<
             ? (
                 await memory.rememberMessages({
                   threadId,
+                  resourceId,
                   config: memoryConfig,
                   vectorMessageSearch: messages
                     .slice(-1)
@@ -481,7 +493,13 @@ export class Agent<
 
     // Get memory tools if available
     const memory = this.getMemory();
-    const memoryTools = memory?.getTools();
+    const memoryTools = memory?.getTools?.();
+
+    let mastraProxy = undefined;
+    const logger = this.logger;
+    if (this.#mastra) {
+      mastraProxy = createMastraProxy({ mastra: this.#mastra, logger });
+    }
 
     const converted = Object.entries(this.tools || {}).reduce(
       (memo, value) => {
@@ -489,46 +507,17 @@ export class Agent<
         const tool = this.tools[k];
 
         if (tool) {
-          memo[k] = {
-            description: tool.description,
-            parameters: tool.inputSchema,
-            execute:
-              typeof tool?.execute === 'function'
-                ? async (args, options) => {
-                    try {
-                      this.logger.debug(`[Agent:${this.name}] - Executing tool ${k}`, {
-                        name: k,
-                        description: tool.description,
-                        args,
-                        runId,
-                        threadId,
-                        resourceId,
-                      });
-                      return (
-                        tool?.execute?.(
-                          {
-                            context: args,
-                            mastra: this.#mastra,
-                            memory,
-                            runId,
-                            threadId,
-                            resourceId,
-                          },
-                          options,
-                        ) ?? undefined
-                      );
-                    } catch (err) {
-                      this.logger.error(`[Agent:${this.name}] - Failed execution`, {
-                        error: err,
-                        runId,
-                        threadId,
-                        resourceId,
-                      });
-                      throw err;
-                    }
-                  }
-                : undefined,
+          const options = {
+            name: k,
+            runId,
+            threadId,
+            resourceId,
+            logger: this.logger,
+            mastra: mastraProxy as MastraUnion | undefined,
+            memory,
+            agentName: this.name,
           };
+          memo[k] = makeCoreTool(tool, options);
         }
         return memo;
       },
@@ -558,7 +547,7 @@ export class Agent<
                           tool?.execute?.(
                             {
                               context: args,
-                              mastra: this.#mastra,
+                              mastra: mastraProxy as MastraUnion | undefined,
                               memory,
                               runId,
                               threadId,
@@ -599,44 +588,15 @@ export class Agent<
       toolsFromToolsets.forEach(toolset => {
         Object.entries(toolset).forEach(([toolName, tool]) => {
           const toolObj = tool;
-          toolsFromToolsetsConverted[toolName] = {
-            description: toolObj.description || '',
-            parameters: toolObj.inputSchema,
-            execute:
-              typeof toolObj?.execute === 'function'
-                ? async (args, options) => {
-                    try {
-                      this.logger.debug(`[Agent:${this.name}] - Executing tool ${toolName}`, {
-                        name: toolName,
-                        description: toolObj.description,
-                        args,
-                        runId,
-                        threadId,
-                        resourceId,
-                      });
-                      return (
-                        toolObj?.execute?.(
-                          {
-                            context: args,
-                            runId,
-                            threadId,
-                            resourceId,
-                          },
-                          options,
-                        ) ?? undefined
-                      );
-                    } catch (error) {
-                      this.logger.error(`[Agent:${this.name}] - Failed toolset execution`, {
-                        error,
-                        runId,
-                        threadId,
-                        resourceId,
-                      });
-                      throw error;
-                    }
-                  }
-                : undefined,
+          const options = {
+            name: toolName,
+            runId,
+            threadId,
+            resourceId,
+            logger: this.logger,
+            agentName: this.name,
           };
+          toolsFromToolsetsConverted[toolName] = makeCoreTool(toolObj, options, 'toolset');
         });
       });
     }