+++ packages/core/src/tools/types.ts
@@ -1,23 +1,31 @@
+import type { ToolExecutionOptions, Tool } from 'ai';
 import type { ZodSchema, z } from 'zod';
 
-import type { IAction, IExecutionContext } from '../action';
-import type { WorkflowContext } from '../workflows';
-import type { ToolExecutionOptions } from 'ai';
+import type { IAction, IExecutionContext, MastraUnion } from '../action';
+import type { Mastra } from '../mastra';
+
+export type VercelTool = Tool;
 
 export type CoreTool = {
   description?: string;
   parameters: ZodSchema;
   execute?: (params: any, options: ToolExecutionOptions) => Promise<any>;
 };
-export interface ToolExecutionContext<
-  TSchemaIn extends z.ZodSchema | undefined = undefined,
-  TContext extends WorkflowContext = WorkflowContext,
-> extends IExecutionContext<TSchemaIn, TContext> {}
+export interface ToolExecutionContext<TSchemaIn extends z.ZodSchema | undefined = undefined>
+  extends IExecutionContext<TSchemaIn> {
+  mastra?: MastraUnion;
+}
 
 export interface ToolAction<
-  TId extends string,
   TSchemaIn extends z.ZodSchema | undefined = undefined,
   TSchemaOut extends z.ZodSchema | undefined = undefined,
   TContext extends ToolExecutionContext<TSchemaIn> = ToolExecutionContext<TSchemaIn>,
   TOptions extends unknown = unknown,
-> extends IAction<TId, TSchemaIn, TSchemaOut, TContext, TOptions> {}
+> extends IAction<string, TSchemaIn, TSchemaOut, TContext, TOptions> {
+  description: string;
+  execute: (
+    context: TContext,
+    options?: TOptions,
+  ) => Promise<TSchemaOut extends z.ZodSchema ? z.infer<TSchemaOut> : unknown>;
+  mastra?: Mastra;
+}