+++ packages/core/src/workflows/machine.ts
@@ -2,19 +2,22 @@ import EventEmitter from 'node:events';
 import type { Span } from '@opentelemetry/api';
 import { get } from 'radash';
 import sift from 'sift';
-import { assign, createActor, fromPromise, setup } from 'xstate';
 import type { MachineContext, Snapshot } from 'xstate';
+import { assign, createActor, fromPromise, setup } from 'xstate';
 import type { z } from 'zod';
 
-import type { IAction, MastraPrimitives } from '../action';
+import type { IAction, MastraUnion } from '../action';
 import type { Logger } from '../logger';
 
+import type { Mastra } from '../mastra';
+import { createMastraProxy } from '../utils';
 import type { Step } from './step';
 import type {
   DependencyCheckOutput,
   ResolverFunctionInput,
   ResolverFunctionOutput,
   RetryConfig,
+  StepAction,
   StepCondition,
   StepDef,
   StepGraph,
@@ -28,6 +31,7 @@ import type {
   WorkflowEvent,
   WorkflowState,
 } from './types';
+import { WhenConditionReturnValue } from './types';
 import {
   getResultActivePaths,
   getStepResult,
@@ -40,10 +44,10 @@ import type { WorkflowInstance } from './workflow-instance';
 
 export class Machine<
   TSteps extends Step<any, any, any>[] = any,
-  TTriggerSchema extends z.ZodType<any> = any,
+  TTriggerSchema extends z.ZodObject<any> = any,
 > extends EventEmitter {
   logger: Logger;
-  #mastra?: MastraPrimitives;
+  #mastra?: Mastra;
   #workflowInstance: WorkflowInstance;
   #executionSpan?: Span | undefined;
 
@@ -54,7 +58,7 @@ export class Machine<
   name: string;
 
   #actor: ReturnType<typeof createActor<ReturnType<typeof this.initializeMachine>>> | null = null;
-  #steps: Record<string, IAction<any, any, any, any>> = {};
+  #steps: Record<string, StepAction<any, any, any, any>> = {};
   #retryConfig?: RetryConfig;
 
   constructor({
@@ -70,7 +74,7 @@ export class Machine<
     startStepId,
   }: {
     logger: Logger;
-    mastra?: MastraPrimitives;
+    mastra?: Mastra;
     workflowInstance: WorkflowInstance;
     executionSpan?: Span;
     name: string;
@@ -319,6 +323,11 @@ export class Machine<
           runId: this.#runId,
         });
 
+        const logger = this.logger;
+        let mastraProxy = undefined;
+        if (this.#mastra) {
+          mastraProxy = createMastraProxy({ mastra: this.#mastra, logger });
+        }
         const result = await stepNode.config.handler({
           context: resolvedData,
           suspend: async (payload?: any) => {
@@ -336,7 +345,7 @@ export class Machine<
             }
           },
           runId: this.#runId,
-          mastra: this.#mastra,
+          mastra: mastraProxy as MastraUnion | undefined,
         });
 
         this.logger.debug(`Step ${stepNode.step.id} result`, {
@@ -386,7 +395,7 @@ export class Machine<
         });
 
         if (typeof stepConfig?.when === 'function') {
-          const conditionMet = await stepConfig.when({
+          let conditionMet = await stepConfig.when({
             context: {
               ...context,
               getStepResult: ((stepId: string) => {
@@ -402,7 +411,12 @@ export class Machine<
             },
             mastra: this.#mastra,
           });
-          if (conditionMet) {
+          if (conditionMet === WhenConditionReturnValue.ABORT) {
+            conditionMet = false;
+          } else if (conditionMet === WhenConditionReturnValue.CONTINUE_FAILED) {
+            // TODO: send another kind of event instead
+            return { type: 'CONDITIONS_SKIPPED' as const };
+          } else if (conditionMet) {
             this.logger.debug(`Condition met for step ${stepNode.step.id}`, {
               stepId: stepNode.step.id,
               runId: this.#runId,
@@ -435,7 +449,11 @@ export class Machine<
         }) => {
           const { parentStepId, context } = input;
           const result = await this.#workflowInstance.runMachine(parentStepId, context);
-          return Promise.resolve({ steps: result?.results });
+          return Promise.resolve({
+            steps: result.reduce((acc, r) => {
+              return { ...acc, ...r?.results };
+            }, {}),
+          });
         },
       ),
     };
@@ -632,6 +650,12 @@ export class Machine<
               },
               {
                 guard: ({ event }: { event: { output: DependencyCheckOutput } }) => {
+                  return event.output.type === 'CONDITIONS_SKIPPED';
+                },
+                target: 'completed',
+              },
+              {
+                guard: ({ event }: { event: { output: DependencyCheckOutput } }) => {
                   return event.output.type === 'CONDITION_FAILED';
                 },
                 target: 'failed',
@@ -815,7 +839,7 @@ export class Machine<
     };
   }
 
-  #evaluateCondition<TStep extends StepVariableType<any, any, any, any>, TTriggerSchema extends z.ZodType<any>>(
+  #evaluateCondition<TStep extends StepVariableType<any, any, any, any>, TTriggerSchema extends z.ZodObject<any>>(
     condition: StepCondition<TStep, TTriggerSchema>,
     context: WorkflowContext,
   ): boolean {
@@ -899,6 +923,14 @@ export class Machine<
       orBranchResult = condition.or.some(cond => this.#evaluateCondition(cond, context));
       this.logger.debug(`Evaluated OR condition`, {
         orBranchResult,
+        runId: this.#runId,
+      });
+    }
+
+    if ('not' in condition) {
+      baseResult = !this.#evaluateCondition(condition.not, context);
+      this.logger.debug(`Evaluated NOT condition`, {
+        baseResult,
         runId: this.#runId,
       });
     }