+++ packages/memory/src/index.ts
@@ -1,11 +1,10 @@
 import { deepMerge } from '@mastra/core';
-import type { CoreMessage, CoreTool } from '@mastra/core';
+import type { AiMessageType, CoreMessage, CoreTool } from '@mastra/core';
 import { MastraMemory } from '@mastra/core/memory';
 import type { MessageType, MemoryConfig, SharedMemoryConfig, StorageThreadType } from '@mastra/core/memory';
 import type { StorageGetMessagesArg } from '@mastra/core/storage';
-import { updateWorkingMemoryTool } from './tools/working-memory';
 import { embed } from 'ai';
-import type { Message as AiMessage } from 'ai';
+import { updateWorkingMemoryTool } from './tools/working-memory';
 
 /**
  * Concrete implementation of MastraMemory that adds support for thread configuration
@@ -24,11 +23,26 @@ export class Memory extends MastraMemory {
     this.threadConfig = mergedConfig;
   }
 
+  private async validateThreadIsOwnedByResource(threadId: string, resourceId: string) {
+    const thread = await this.storage.getThreadById({ threadId });
+    if (!thread) {
+      throw new Error(`No thread found with id ${threadId}`);
+    }
+    if (thread.resourceId !== resourceId) {
+      throw new Error(
+        `Thread with id ${threadId} is for resource with id ${thread.resourceId} but resource ${resourceId} was queried.`,
+      );
+    }
+  }
+
   async query({
     threadId,
+    resourceId,
     selectBy,
     threadConfig,
-  }: StorageGetMessagesArg): Promise<{ messages: CoreMessage[]; uiMessages: AiMessage[] }> {
+  }: StorageGetMessagesArg): Promise<{ messages: CoreMessage[]; uiMessages: AiMessageType[] }> {
+    if (resourceId) await this.validateThreadIsOwnedByResource(threadId, resourceId);
+
     let vectorResults:
       | null
       | {
@@ -108,20 +122,29 @@ export class Memory extends MastraMemory {
 
   async rememberMessages({
     threadId,
+    resourceId,
     vectorMessageSearch,
     config,
   }: {
     threadId: string;
+    resourceId?: string;
     vectorMessageSearch?: string;
     config?: MemoryConfig;
-  }) {
+  }): Promise<{
+    threadId: string;
+    messages: CoreMessage[];
+    uiMessages: AiMessageType[];
+  }> {
+    if (resourceId) await this.validateThreadIsOwnedByResource(threadId, resourceId);
+
     const threadConfig = this.getMergedThreadConfig(config || {});
 
     if (!threadConfig.lastMessages && !threadConfig.semanticRecall) {
       return {
         messages: [],
         uiMessages: [],
-      } satisfies Awaited<ReturnType<typeof this.query>>;
+        threadId,
+      };
     }
 
     const messages = await this.query({
@@ -134,7 +157,11 @@ export class Memory extends MastraMemory {
     });
 
     this.logger.debug(`Remembered message history includes ${messages.messages.length} messages.`);
-    return messages;
+    return {
+      threadId,
+      messages: messages.messages,
+      uiMessages: messages.uiMessages,
+    };
   }
 
   async getThreadById({ threadId }: { threadId: string }): Promise<StorageThreadType | null> {